ANDROID: arm64: add support for building the KASLR kernel with LLVM lld
Move the definition of some __efistub_ decorated section markers to the
linker script, which permits us to assign them as relative quantities
(since using an intermediate assignment loses the section relative
property when using ld.lld)

This was split off of the patch originally written by Ard Biesheuvel.
It's not clear if this is better fixed by support for forward section
declarations in LLD, or reordering of the sections as we've done here.

Bug: 63740206
Bug: 120083959
Change-Id: I9bd77de9bac7c970e6082e90302a6eacd82a7ce7
Link: https://www.spinics.net/lists/arm-kernel/msg691488.html
Link: https://github.com/ClangBuiltLinux/linux/issues/280
Sugguested-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>